Cardano.Ledger.Binary

Documentation

module Cardano.Ledger.Binary.Decoding

module Cardano.Ledger.Binary.Encoding

module Cardano.Ledger.Binary.Group

module Cardano.Ledger.Binary.Version

class Typeable a => FromCBOR a where

Methods

fromCBOR :: Decoder s a

class Typeable a => ToCBOR a where

Methods

toCBOR :: a -> Encoding

data Term

Constructors

TInt !Int

TInteger !Integer

TBytes !ByteString

TBytesI !ByteString

TString !Text

TStringI !Text

TList ![Term]

TListI ![Term]

TMap ![(Term, Term)]

TMapI ![(Term, Term)]

TTagged !Word64 !Term

TBool !Bool

TNull

TSimple !Word8

THalf !Float

TFloat !Float

TDouble !Double

data DeserialiseFailure

Constructors

DeserialiseFailure ByteOffset String

translateViaCBORAnnotator

Arguments

:: (ToCBOR a, DecCBOR (Annotator b))

=> Version
Version that will be used for deserialization

-> Text

-> a

-> Except DecoderError b

Translation function between values through a related binary representation. This function allows you to translate one type into another (or the same one) through their common binary format. It is possible for the source type to be encoded with a different version than the version that will be used for decoding. This is useful for types that build upon one another and are "upgradeable" through their binary representation. It is important to note that the deserialization will happen with Annotator, since that is usually the way we deserialize upgradeable types that live on chain. Moreover, encoding does not require a version, because memoized types that were decoded with annotation will have the bytes retained and thus will have the ToCBOR instance.