Cardano.Ledger.Binary.Decoding.Decoder

Decoders

data Decoder s a

toPlainDecoder ∷ Version → Decoder s a → Decoder s a

Extract the underlying Decoder by specifying the concrete version to be used.

fromPlainDecoder ∷ Decoder s a → Decoder s a

Promote a regular Decoder to a versioned one. Which means it will work for all versions.

withPlainDecoder ∷ Decoder s a → (Decoder s a → Decoder s b) → Decoder s b

Use the supplied decoder as a plain decoder with current version.

enforceDecoderVersion ∷ Version → Decoder s a → Decoder s a

Ignore the current version of the decoder and enforce the supplied one instead.

data DecoderError

Constructors

DecoderErrorCanonicityViolation Text

DecoderErrorCustom Text Text
Custom decoding error, usually due to some validation failure

DecoderErrorDeserialiseFailure Text DeserialiseFailure

DecoderErrorEmptyList Text

DecoderErrorLeftover Text ByteString

DecoderErrorSizeMismatch Text Int Int
A size mismatch DecoderErrorSizeMismatch label expectedSize actualSize

DecoderErrorUnknownTag Text Word8

DecoderErrorVoid

type ByteOffset = Int64

A 0-based offset within the overall byte sequence that makes up the input to the Decoder.

This is an Int64 since Decoder is incremental and can decode more data than fits in memory at once. This is also compatible with the result type of length.

data DecodeAction s a

An action, representing a step for a decoder to taken and a continuation to invoke with the expected value.

Since: cborg-0.2.0.0

Constructors

ConsumeWord (Word# → ST s (DecodeAction s a))

ConsumeWord8 (Word# → ST s (DecodeAction s a))

ConsumeWord16 (Word# → ST s (DecodeAction s a))

ConsumeWord32 (Word# → ST s (DecodeAction s a))

ConsumeNegWord (Word# → ST s (DecodeAction s a))

ConsumeInt (Int# → ST s (DecodeAction s a))

ConsumeInt8 (Int# → ST s (DecodeAction s a))

ConsumeInt16 (Int# → ST s (DecodeAction s a))

ConsumeInt32 (Int# → ST s (DecodeAction s a))

ConsumeListLen (Int# → ST s (DecodeAction s a))

ConsumeMapLen (Int# → ST s (DecodeAction s a))

ConsumeTag (Word# → ST s (DecodeAction s a))

ConsumeInteger (Integer → ST s (DecodeAction s a))

ConsumeFloat (Float# → ST s (DecodeAction s a))

ConsumeDouble (Double# → ST s (DecodeAction s a))

ConsumeBytes (ByteString → ST s (DecodeAction s a))

ConsumeByteArray (ByteArray → ST s (DecodeAction s a))

ConsumeString (Text → ST s (DecodeAction s a))

ConsumeUtf8ByteArray (ByteArray → ST s (DecodeAction s a))

ConsumeBool (Bool → ST s (DecodeAction s a))

ConsumeSimple (Word# → ST s (DecodeAction s a))

ConsumeBytesIndef (ST s (DecodeAction s a))

ConsumeStringIndef (ST s (DecodeAction s a))

ConsumeListLenIndef (ST s (DecodeAction s a))

ConsumeMapLenIndef (ST s (DecodeAction s a))

ConsumeNull (ST s (DecodeAction s a))

ConsumeListLenOrIndef (Int# → ST s (DecodeAction s a))

ConsumeMapLenOrIndef (Int# → ST s (DecodeAction s a))

ConsumeBreakOr (Bool → ST s (DecodeAction s a))

PeekTokenType (TokenType → ST s (DecodeAction s a))

PeekAvailable (Int# → ST s (DecodeAction s a))

PeekByteOffset (Int# → ST s (DecodeAction s a))

ConsumeWordCanonical (Word# → ST s (DecodeAction s a))

ConsumeWord8Canonical (Word# → ST s (DecodeAction s a))

ConsumeWord16Canonical (Word# → ST s (DecodeAction s a))

ConsumeWord32Canonical (Word# → ST s (DecodeAction s a))

ConsumeNegWordCanonical (Word# → ST s (DecodeAction s a))

ConsumeIntCanonical (Int# → ST s (DecodeAction s a))

ConsumeInt8Canonical (Int# → ST s (DecodeAction s a))

ConsumeInt16Canonical (Int# → ST s (DecodeAction s a))

ConsumeInt32Canonical (Int# → ST s (DecodeAction s a))

ConsumeListLenCanonical (Int# → ST s (DecodeAction s a))

ConsumeMapLenCanonical (Int# → ST s (DecodeAction s a))

ConsumeTagCanonical (Word# → ST s (DecodeAction s a))

ConsumeIntegerCanonical (Integer → ST s (DecodeAction s a))

ConsumeFloat16Canonical (Float# → ST s (DecodeAction s a))

ConsumeFloatCanonical (Float# → ST s (DecodeAction s a))

ConsumeDoubleCanonical (Double# → ST s (DecodeAction s a))

ConsumeBytesCanonical (ByteString → ST s (DecodeAction s a))

ConsumeByteArrayCanonical (ByteArray → ST s (DecodeAction s a))

ConsumeStringCanonical (Text → ST s (DecodeAction s a))

ConsumeUtf8ByteArrayCanonical (ByteArray → ST s (DecodeAction s a))

ConsumeSimpleCanonical (Word# → ST s (DecodeAction s a))

Fail String

Done a

data TokenType

The type of a token, which a decoder can ask for at an arbitrary time.

Since: cborg-0.2.0.0

Constructors

TypeUInt

TypeUInt64

TypeNInt

TypeNInt64

TypeInteger

TypeFloat16

TypeFloat32

TypeFloat64

TypeBytes

TypeBytesIndef

TypeString

TypeStringIndef

TypeListLen

TypeListLen64

TypeListLenIndef

TypeMapLen

TypeMapLen64

TypeMapLenIndef

TypeTag

TypeTag64

TypeBool

TypeNull

TypeSimple

TypeBreak

TypeInvalid

Versioning

getDecoderVersion ∷ Decoder s Version

Extract current version of the decoder

>>> import Cardano.Ledger.Decoding
>>> decodeFullDecoder 3 "Version" getDecoderVersion ""
Right 3

ifDecoderVersionAtLeast

Arguments

∷ Version

→ Decoder s a
Use this decoder if current decoder version is larger or equal to the supplied Version

→ Decoder s a
Use this decoder if current decoder version is lower than the supplied Version

→ Decoder s a

Conditionally choose the newer or older decoder, depending on the current version. Version in the context of encoders/decoders is the major protocol version. Supplied version acts as a pivot.

Example

Expand

Let's say prior to the version 2 some type Foo was backed by Word16, but at the 2nd version onwards it was switched to Word32 instead. In order to support both versions, we change the type, but we also use this condition to keep backwards compatibility of the decoder:

>>> newtype Foo = Foo Word32
>>> decFoo = Foo <$> ifDecoderVersionAtLeast 2 decodeWord32 (fromIntegral <$> decodeWord16)

whenDecoderVersionAtLeast

Arguments

∷ Version

→ Decoder s a
Run this decoder whenever current decoder version is larger or equal to the supplied Version

→ Decoder s ()

Optionally run a decoder depending on the current version and the supplied one.

unlessDecoderVersionAtLeast

Arguments

∷ Version

→ Decoder s a
Run this decoder whenever current decoder version is smaller to the supplied Version

→ Decoder s ()

Optionally run a decoder depending on the current version and the supplied one.

Error reporting

cborError ∷ (MonadFail m, Buildable e) ⇒ e → m a

Convert a Buildable error message into a MonadFail failure.

toCborError ∷ (MonadFail m, Buildable e) ⇒ Either e a → m a

Convert an Either-encoded failure to a MonadFail failure using the Buildable insatance

showDecoderError ∷ Buildable e ⇒ e → String

invalidKey ∷ MonadFail m ⇒ Word → m a

Report an error when a numeric key of the type constructor doesn't match.

assertTag ∷ Word → Decoder s ()

enforceSize ∷ Text → Int → Decoder s ()

Enforces that the input size is the same as the decoded one, failing in case it's not

matchSize ∷ Text → Int → Int → Decoder s ()

Compare two sizes, failing if they are not equal

Compatibility tools

binaryGetDecoder

Arguments

∷ Bool
Flag to allow left over at the end or not

→ Text
Name of the function or type for error reporting

→ Get a
Deserializer for the binary package

→ Decoder s a

Convert a Get monad from binary package into a Decoder

allowTag ∷ Word → Decoder s ()

Custom decoders

decodeVersion ∷ Decoder s Version

decodeRational ∷ Decoder s Rational

Decoder for Rational. Versions variance:

[>= 9] - Allows variable as well as exact list length encoding. Consumes tag 30 if one is present, but does not enforce it.

[>= 2] - Allows variable as well as exact list length encoding.

[== 1] - Expects exact list length encoding.

decodeRationalWithTag ∷ Decoder s Rational

Enforces tag 30 to indicate a rational number, as per tag assignment: https://www.iana.org/assignments/cbor-tags/cbor-tags.xhtml

https://peteroupc.github.io/CBOR/rational.html

decodeRecordNamed ∷ Text → (a → Int) → Decoder s a → Decoder s a

decodeRecordNamedT ∷ (MonadTrans m, Monad (m (Decoder s))) ⇒ Text → (a → Int) → m (Decoder s) a → m (Decoder s) a

decodeRecordSum ∷ Text → (Word → Decoder s (Int, a)) → Decoder s a

decodeListLike

Arguments

∷ Text
Name for error reporting

→ Decoder s a
Decoder for the datastructure itself

→ (a → Int → Decoder s ())
In case when length was encoded, act upon it.

→ Decoder s a

Use this decoder for any list like structure that accepts fixed or variable list length encoding.

decodeListLikeT

Arguments

∷ (MonadTrans m, Monad (m (Decoder s)))

⇒ Text
Name for error reporting

→ m (Decoder s) a
Decoder for the datastructure itself

→ (a → Int → m (Decoder s) ())
In case when length was encoded, act upon it.

→ m (Decoder s) a

decodeEnumBounded ∷ ∀ a s. (Enum a, Bounded a, Typeable a) ⇒ Decoder s a

decodeWithOrigin ∷ Decoder s a → Decoder s (WithOrigin a)

Containers

decodeMaybe ∷ Decoder s a → Decoder s (Maybe a)

Decoder for Maybe. Versions variance:

[>= 2] - Allows variable as well as exact list length encoding.

[< 2] - Expects exact list length encoding

decodeNullMaybe ∷ Decoder s a → Decoder s (Maybe a)

Alternative way to decode a Maybe type.

Note - this is not the default method for decoding Maybe, use decodeMaybe instead.

decodeStrictMaybe ∷ Decoder s a → Decoder s (StrictMaybe a)

Unlike decodeMaybe this allows variable as well as exact list length encoding for all versions, because Byron never used StrictMaybe type.

decodeNullStrictMaybe ∷ Decoder s a → Decoder s (StrictMaybe a)

Alternative way to decode a StrictMaybe type.

Note - this is not the default method for decoding StrictMaybe, use decodeStrictMaybe instead.

decodeEither ∷ Decoder s a → Decoder s b → Decoder s (Either a b)

decodeList ∷ Decoder s a → Decoder s [a]

Decoder for list.

[>= 2] - Allows variable as well as exact list length encoding.

[< 2] - Expects variable list length encoding

decodeNonEmptyList ∷ Decoder s a → Decoder s (NonEmpty a)

decodeVector ∷ Vector vec a ⇒ Decoder s a → Decoder s (vec a)

Generic decoder for vectors. Its intended use is to allow easy definition of Serialise instances for custom vector

decodeSet ∷ Ord a ⇒ Decoder s a → Decoder s (Set a)

Decoder for Set. Versions variance:

[>= 9] - Allows variable as well as exact list length encoding. Duplicates are not allowed. Set tag 258 is permitted, but not enforced.

[>= 2, < 9] - Allows variable as well as exact list length encoding. Duplicates are silently ignored, set tag 258 is not permitted.

[< 2] - Expects exact list length encoding and enforces strict order without any duplicates. Also enforces special set tag 258, which was abandoned starting with version 2

setTag ∷ Word

We stitch a `258` in from of a (Hash)Set, so that tools which programmatically check for canonicity can recognise it from a normal array. Why 258? This will be formalised pretty soon, but IANA allocated 256...18446744073709551615 to "First come, first served": https://www.iana.org/assignments/cbor-tags/cbor-tags.xhtml Currently `258` is the first unassigned tag and as it requires 2 bytes to be encoded, it sounds like the best fit.

https://github.com/input-output-hk/cbor-sets-spec/blob/master/CBOR_SETS.md

decodeMap ∷ Ord k ⇒ Decoder s k → Decoder s v → Decoder s (Map k v)

Decoder for Map. Versions variance:

[>= 9] - Allows variable as well as exact list length encoding. Duplicate keys will result in a deserialization failure

[>= 2] - Allows variable as well as exact list length encoding. Duplicate keys are silently ignored

[< 2] - Expects exact list length encoding and enforces strict order without any duplicates.

An example of how to use versioning

>>> :set -XOverloadedStrings
>>> import Codec.CBOR.FlatTerm
>>> fromFlatTerm (toPlainDecoder 1 (decodeMap decodeInt decodeBytes)) [TkMapLen 2,TkInt 1,TkBytes "Foo",TkInt 2,TkBytes "Bar"]
Right (fromList [(1,"Foo"),(2,"Bar")])
>>> fromFlatTerm (toPlainDecoder 1 (decodeMap decodeInt decodeBytes)) [TkMapBegin,TkInt 1,TkBytes "Foo",TkInt 2,TkBytes "Bar"]
Left "decodeMapLen: unexpected token TkMapBegin"
>>> fromFlatTerm (toPlainDecoder 2 (decodeMap decodeInt decodeBytes)) [TkMapBegin,TkInt 1,TkBytes "Foo",TkInt 2,TkBytes "Bar",TkBreak]
Right (fromList [(1,"Foo"),(2,"Bar")])

decodeMapByKey ∷ Ord k ⇒ Decoder s k → (k → Decoder s v) → Decoder s (Map k v)

Just like decodeMap, but also gives access to the key for the value decoder.

decodeMapLikeEnforceNoDuplicates ∷ Ord k ⇒ Decoder s (Maybe Int) → Decoder s (k, v) → Decoder s (Map k v)

Similar to decodeMapByKey, except it gives access to the key value decoder as a pair and allows for different type of length encoding

decodeVMap ∷ (Vector kv k, Vector vv v, Ord k) ⇒ Decoder s k → Decoder s v → Decoder s (VMap kv vv k v)

Decode VMap. Unlike decodeMap it does not behavee differently for version prior to 2.

decodeSeq ∷ Decoder s a → Decoder s (Seq a)

Decoder for Seq. Same behavior for all versions, allows variable as well as exact list length encoding

decodeStrictSeq ∷ Decoder s a → Decoder s (StrictSeq a)

Decoder for StrictSeq. Same behavior for all versions, allows variable as well as exact list length encoding.

decodeSetTag ∷ Decoder s ()

decodeListLikeWithCount

Arguments

∷ ∀ s a b. Monoid b

⇒ Decoder s (Maybe Int)
Length decoder that produces the expected number of elements. When Nothing is decoded the decodeBreakOr will be used as termination indicator.

→ (a → b → b)
Add an element into the decoded List like data structure

→ (b → Decoder s a)
Decoder for the values. Current accumulator is supplied as an argument

→ Decoder s (Int, b)

Decode a collection of values with ability to supply length decoder. Number of decoded elements will be returned together with the data structure

decodeSetLikeEnforceNoDuplicates

Arguments

∷ ∀ s a b c. Monoid b

⇒ (a → b → b)
Add an element into the decoded Set like data structure

→ (b → (Int, c))
Get the final data structure from the decoded sequence of values and the number of elements it contains. This is useful when a sequence on the wire is represented by a set, namely no duplicates are allowed, while the Haskell representation uses some other data structure that enforces no duplicates by some other means. For example a Map, where keys are hashes of the values encoded on the wire. The size of the final data structure will be used to enforce the invariant that the number of elements decoded matches the final size of the Set like data structure, thus ensuring no duplicates were encountered.

→ Decoder s a

→ Decoder s c

Decode a Set as a either a definite or indefinite list. Duplicates are not allowed. Set tag 258 is permitted, but not enforced.

decodeListLikeEnforceNoDuplicates

Arguments

∷ ∀ s a b c. Monoid b

⇒ Decoder s (Maybe Int)

→ (a → b → b)
Add an element into the decoded List like data structure

→ (b → (Int, c))
Get the final data structure and the number of elements it has.

→ Decoder s a

→ Decoder s c

Decode a collection of values with ability to supply length decoder. Duplicates are not allowed.

decodeMapContents ∷ Decoder s a → Decoder s [a]

Applicaitve

decodeMapTraverse ∷ (Ord a, Applicative t) ⇒ Decoder s (t a) → Decoder s (t b) → Decoder s (t (Map a b))

decodeMapContentsTraverse ∷ Applicative t ⇒ Decoder s (t a) → Decoder s (t b) → Decoder s (t [(a, b)])

Time

decodeUTCTime ∷ Decoder s UTCTime

Decoder for UTCTime. Versions variance:

[>= 2] - Allows variable list length encoding, but still expects number of elements to be 3.

[< 2] - Expects exact list length encoding to be 3

Network

decodeIPv4 ∷ Decoder s IPv4

decodeIPv6 ∷ Decoder s IPv6

Lifted cborg decoders

decodeBool ∷ Decoder s Bool

decodeBreakOr ∷ Decoder s Bool

decodeByteArray ∷ Decoder s ByteArray

decodeByteArrayCanonical ∷ Decoder s ByteArray

decodeBytes ∷ Decoder s ByteString

decodeBytesCanonical ∷ Decoder s ByteString

decodeBytesIndef ∷ Decoder s ()

decodeDouble ∷ Decoder s Double

decodeDoubleCanonical ∷ Decoder s Double

decodeFloat ∷ Decoder s Float

decodeFloat16Canonical ∷ Decoder s Float

decodeFloatCanonical ∷ Decoder s Float

decodeInt ∷ Decoder s Int

decodeInt16 ∷ Decoder s Int16

decodeInt16Canonical ∷ Decoder s Int16

decodeInt32 ∷ Decoder s Int32

decodeInt32Canonical ∷ Decoder s Int32

decodeInt64 ∷ Decoder s Int64

decodeInt64Canonical ∷ Decoder s Int64

decodeInt8 ∷ Decoder s Int8

decodeInt8Canonical ∷ Decoder s Int8

decodeIntCanonical ∷ Decoder s Int

decodeInteger ∷ Decoder s Integer

decodeIntegerCanonical ∷ Decoder s Integer

decodeNatural ∷ Decoder s Natural

decodeListLen ∷ Decoder s Int

decodeListLenCanonical ∷ Decoder s Int

decodeListLenCanonicalOf ∷ Int → Decoder s ()

decodeListLenIndef ∷ Decoder s ()

decodeListLenOf ∷ Int → Decoder s ()

decodeListLenOrIndef ∷ Decoder s (Maybe Int)

decodeMapLen ∷ Decoder s Int

decodeMapLenCanonical ∷ Decoder s Int

decodeMapLenIndef ∷ Decoder s ()

decodeMapLenOrIndef ∷ Decoder s (Maybe Int)

decodeNegWord ∷ Decoder s Word

decodeNegWord64 ∷ Decoder s Word64

decodeNegWord64Canonical ∷ Decoder s Word64

decodeNegWordCanonical ∷ Decoder s Word

decodeNull ∷ Decoder s ()

decodeSequenceLenIndef ∷ (r → a → r) → r → (r → b) → Decoder s a → Decoder s b

decodeSequenceLenN ∷ (r → a → r) → r → (r → b) → Int → Decoder s a → Decoder s b

decodeSimple ∷ Decoder s Word8

decodeSimpleCanonical ∷ Decoder s Word8

decodeString ∷ Decoder s Text

decodeStringCanonical ∷ Decoder s Text

decodeStringIndef ∷ Decoder s ()

decodeTag ∷ Decoder s Word

decodeTag64 ∷ Decoder s Word64

decodeTag64Canonical ∷ Decoder s Word64

decodeTagCanonical ∷ Decoder s Word

decodeUtf8ByteArray ∷ Decoder s ByteArray

decodeUtf8ByteArrayCanonical ∷ Decoder s ByteArray

decodeWithByteSpan ∷ Decoder s a → Decoder s (a, ByteOffset, ByteOffset)

decodeWord ∷ Decoder s Word

decodeWord16 ∷ Decoder s Word16

decodeWord16Canonical ∷ Decoder s Word16

decodeWord32 ∷ Decoder s Word32

decodeWord32Canonical ∷ Decoder s Word32

decodeWord64 ∷ Decoder s Word64

decodeWord64Canonical ∷ Decoder s Word64

decodeWord8 ∷ Decoder s Word8

decodeWord8Canonical ∷ Decoder s Word8

decodeWordCanonical ∷ Decoder s Word

decodeWordCanonicalOf ∷ Word → Decoder s ()

decodeWordOf ∷ Word → Decoder s ()

decodeTerm ∷ Decoder s Term

peekAvailable ∷ Decoder s Int

peekByteOffset ∷ Decoder s ByteOffset

peekTokenType ∷ Decoder s TokenType