Cardano.Ledger.Binary.Decoding.Drop

Description

DEPRECATED - This module is deprecated and used in Byron era exclusively. For newer implementation of this functionality use dropCBOR

A 'Dropper s' is a 'Decoder s ()', that is a decoder that returns nothing

We use Droppers when we don't care about the result of decoding, for example when we have deprecated some part of the serialised blockchain, but still need to decode old blocks.

Documentation

type Dropper s = Decoder s ()

dropBytes ∷ Dropper s

dropInt32 ∷ Dropper s

dropList ∷ Dropper s → Dropper s

Drop a list of values using the supplied Dropper for each element

dropMap ∷ Dropper s → Dropper s → Dropper s

dropSet ∷ Dropper s → Dropper s

dropTuple ∷ Dropper s → Dropper s → Dropper s

dropTriple ∷ Dropper s → Dropper s → Dropper s → Dropper s

dropWord8 ∷ Dropper s

dropWord64 ∷ Dropper s