Cardano.Ledger.Binary.Decoding.Sharing

Documentation

class Monoid (Share a) ⇒ DecShareCBOR a where

Minimal complete definition

(decShareCBOR | decSharePlusCBOR)

Associated Types

type Share a ∷ Type

type Share a = ()

Methods

getShare ∷ a → Share a

Whenever fromShareCBOR is being used for defining the instance this function should return the state that can be added whenever user invokes decSharePlusCBOR. mempty is returned by default.

decShareCBOR ∷ Share a → Decoder s a

Utilize sharing when decoding, but do not add anything to the state for future sharing.

decSharePlusCBOR ∷ StateT (Share a) (Decoder s) a

Deserialize with sharing and add to the state that is used for sharing. Default implementation will add value returned by getShare for adding to the state.

newtype Interns a

Constructors

Interns [Intern a]

data Intern a

This is an abstract interface that does the interning. In other words it does the actual sharing by looking up the supplied value in some existing data structure and uses that value instead. Relying on this interface gives us the benefit of ignoring the type of underlying data structure and allows us to compose many Interns with the monoidal interface provided by Interns wrapper. In order to create an Intern see the internsFromMap or internsFromVMap functions.

Constructors

Intern

Fields

internMaybe ∷ a → Maybe a

Function that will do the interning. If value is not available then Nothing is returned.

internWeight ∷ !Int

Used for sorting. Normally set to the size of the underlying data structure. Keeping interns sorted with respect to how many elements is in the underlying data structure in theory gives a better chance of successful intern hit sooner rather than later.

decShareLensCBOR ∷ DecShareCBOR b ⇒ SimpleGetter bs (Share b) → StateT bs (Decoder s) b

decSharePlusLensCBOR ∷ DecShareCBOR b ⇒ Lens' bs (Share b) → StateT bs (Decoder s) b

Just like decSharePlusCBOR, except allows to transform the shared state with a lens.

decNoShareCBOR ∷ DecShareCBOR a ⇒ Decoder s a

Use DecShareCBOR class while ignoring sharing

interns ∷ Interns k → k → k

internsFromMap ∷ Ord k ⇒ Map k a → Interns k

internsFromVMap ∷ Ord k ⇒ VMap VB kv k a → Interns k

toMemptyLens ∷ Monoid a ⇒ Lens' a b → Lens' c b → Lens' c a

Using this function it is possible to compose two lenses. One will extract a value and another will used it for placing it into a empty monoid. Here is an example of how a second element of a tuple can be projected on the third element of a 3-tuple.

toMemptyLens _3 _2 == lens (\(_, b) -> (mempty, mempty, b)) (\(a, _) (_, _, b) -> (a, b))

Here is an example where we extract a second element of a tuple and insert it at third position of a three tuple while all other elements are set to mempty:

>>> import Lens.Micro
>>> ("foo","bar") ^. toMemptyLens _3 _2 :: (Maybe String, (), String)
(Nothing,(),"bar")

In the opposite direction of extracting the third element of a 3-tuple and replacing the second element of the tuple the setter is being applied to

>>> ("foo","bar") & toMemptyLens _3 _2 .~ (Just "baz", (), "booyah") :: (String, String)
("foo","booyah")

decShareMonadCBOR ∷ (DecCBOR (f b), Monad f) ⇒ Interns b → Decoder s (f b)

Share every item in a functor, have deserializing it