Cardano.Ledger.Binary.Decoding.Sized

Documentation

data Sized a

A CBOR deserialized value together with its size. When deserializing use either decodeSized or its DecCBOR instance.

Use mkSized to construct such value.

Constructors

Sized

Fields

sizedValue ∷ !a

sizedSize ∷ Int64

Overhead in bytes. The field is lazy on purpose, because it might not be needed, but it can be expensive to compute.

mkSized ∷ EncCBOR a ⇒ Version → a → Sized a

Construct a Sized value by serializing it first and recording the amount of bytes it requires. Note, however, CBOR serialization is not canonical, therefore it is *NOT* a requirement that this property holds:

sizedSize (mkSized a) === sizedSize (unsafeDeserialize (serialize a) :: a)

decodeSized ∷ Decoder s a → Decoder s (Sized a)

sizedDecoder ∷ Decoder s a → Decoder s (Sized a)

Deprecated: In favor of more consistently named decodeSized

toSizedL ∷ EncCBOR s ⇒ Version → Lens' s a → Lens' (Sized s) a

Take a lens that operates on a particular type and convert it into a lens that operates on the Sized version of the type.