Cardano.Ledger.Binary.FlatTerm

Documentation

type FlatTerm = [TermToken]

A "flat" representation of an Encoding value, useful for round-tripping and writing tests.

Since: cborg-0.2.0.0

data TermToken

A concrete encoding of Encoding values, one which mirrors the original Encoding type closely.

Since: cborg-0.2.0.0

Constructors

TkInt !Codec.CBOR.FlatTerm

Methods

compare :: TermToken -> TermToken -> Ordering

(<) :: TermToken -> TermToken -> Bool

(<=) :: TermToken -> TermToken -> Bool

(>) :: TermToken -> TermToken -> Bool

(>=) :: TermToken -> TermToken -> Bool

max :: TermToken -> TermToken -> TermToken

min :: TermToken -> TermToken -> TermToken

toFlatTerm :: Version -> Encoding -> FlatTerm

fromFlatTerm :: Version -> (forall s. Decoder s a) -> FlatTerm -> Either String a

validFlatTerm

Arguments

:: FlatTerm
The input FlatTerm

-> Bool
True if valid, False otherwise.

Ensure a FlatTerm is internally consistent and was created in a valid manner.

Since: cborg-0.2.0.0