Cardano.Ledger.Binary.Version

Versioning

data Version

Protocol version number that is used during encoding and decoding. All supported versions are in the range from MinVersion to MaxVersion.

getVersion :: Integral i => Version -> i

Convert a Version to an Integral value.

Note - Version spans a fairly small range of non-negative numbers, so this should be safe even for smallest integral types.

type MinVersion = 0

Minimum supported version

type MaxVersion = 10

Maximum supported version. This is the protocol version of the next upcoming era

natVersion :: forall v. (KnownNat v, MinVersion <= v, v <= MaxVersion) => Version

Same as natVersionProxy, construct a version from a type level Nat, except it can be supplied through TypeApplications.

natVersionProxy :: (KnownNat v, MinVersion <= v, v <= MaxVersion) => Proxy v -> Version

Safely construct a Version from a type level Nat, which is supplied as a Proxy

succVersion :: MonadFail m => Version -> m Version

Increment version by 1.

mkVersion :: (Integral i, MonadFail m) => i -> m Version

Construct a Version and fail if the supplied value is not a supported version number.

mkVersion64 :: MonadFail m => Word64 -> m Version

Construct a Version and fail if the supplied value is not supported version number.

getVersion64 :: Version -> Word64

Extract Word64 representation of the Version

allVersions :: [Version]

Concrete era versions

byronProtVer :: Version

shelleyProtVer :: Version