Test.Cardano.Ledger.Binary.Plain.RoundTrip

Description

Defines reusable abstractions for testing RoundTrip properties of plain encoders/decoders

Documentation

roundTripSpec ∷ ∀ t. (Show t, Eq t, Typeable t, Arbitrary t) ⇒ Trip t t → Spec

Tests the roundtrip property using QuickCheck generators for all possible versions starting with shelleyProtVer.

roundTripFailureExpectation ∷ ∀ a. (Eq a, ToCBOR a, FromCBOR a) ⇒ a → Expectation

roundTripExpectation ∷ (Show t, Eq t, Typeable t, HasCallStack) ⇒ Trip t t → t → Expectation

Verify that round triping through the binary form holds

roundTripCborExpectation ∷ ∀ t. (Show t, Eq t, FromCBOR t, ToCBOR t, HasCallStack) ⇒ t → Expectation

embedTripSpec ∷ ∀ a b. (Eq b, Show a, Typeable a, Typeable b, Arbitrary a, HasCallStack) ⇒ Trip a b → (b → a → Expectation) → Spec

Tests the embedtrip property using QuickCheck generators

embedTripExpectation ∷ ∀ a b. (Typeable b, Eq b, HasCallStack) ⇒ Trip a b → (b → a → Expectation) → a → Expectation

data RoundTripFailure

Constructors

RoundTripFailure

Fields

rtfEncoding ∷ Encoding

Produced plain encoding

rtfEncodedBytes ∷ ByteString

Serialized encoding

rtfReEncodedBytes ∷ Maybe ByteString

Re-serialized bytes, if there was a mismatch between the binary form and the reserialization of the data type.

rtfFlatTermError ∷ Maybe String

Roundtripping through FlatTerm

rtfDecoderError ∷ Maybe DecoderError

Error received while decoding the produced bytes.

showMaybeDecoderError ∷ Buildable b ⇒ String → Maybe b → String

showFailedTermsWithReSerialization ∷ ByteString → Maybe ByteString → [String]

data Trip a b

A definition of a CBOR trip through binary representation of one type to another. In this module this is called an embed. When a source and target type is the exact same one then it would be a dual and is expected to round trip.

Constructors

Trip

Fields

tripEncoder ∷ a → Encoding

tripDecoder ∷ ∀ s. Decoder s b

mkTrip ∷ ∀ a b. (a → Encoding) → (∀ s. Decoder s b) → Trip a b

Construct a Trip using encoder and decoder, with dropper set to the decoder which drops the value

cborTrip ∷ ∀ a b. (ToCBOR a, FromCBOR b) ⇒ Trip a b

roundTrip ∷ ∀ t. (Eq t, Typeable t) ⇒ Trip t t → t → Either RoundTripFailure t

embedTrip ∷ ∀ a b. (Eq b, Typeable b) ⇒ Trip a b → a → Either RoundTripFailure b

Can we serialise a type, and then deserialise it as something else?

embedTripLabel ∷ ∀ a b. Eq b ⇒ Text → Trip a b → a → Either RoundTripFailure b