Test.Cardano.Ledger.Binary.RoundTrip

Description

Defines reusable abstractions for testing RoundTrip properties of CBOR instances

Spec

roundTripSpec ∷ ∀ t. (Show t, Eq t, Typeable t, Arbitrary t) ⇒ Trip t t → Spec

Tests the roundtrip property using QuickCheck generators for all possible versions starting with shelleyProtVer.

roundTripCborSpec ∷ ∀ t. (Show t, Eq t, Arbitrary t, EncCBOR t, DecCBOR t) ⇒ Spec

Tests the roundtrip property using QuickCheck generators for all possible versions starting with shelleyProtVer.

roundTripAnnCborSpec ∷ ∀ t. (Show t, Eq t, Arbitrary t, ToCBOR t, DecCBOR (Annotator t)) ⇒ Spec

Tests the roundtrip property using QuickCheck generators for all possible versions starting with shelleyProtVer.

roundTripRangeSpec ∷ ∀ t. (Show t, Eq t, Typeable t, Arbitrary t) ⇒ Trip t t → Version → Version → Spec

Tests the roundtrip property using QuickCheck generators for specific range of versions

Expectations

Trip

roundTripExpectation ∷ (Show t, Eq t, Typeable t, HasCallStack) ⇒ Trip t t → t → Expectation

Verify that round triping through the binary form holds for all versions starting with shelleyProtVer.

roundTripRangeExpectation

Arguments

∷ ∀ t. (Show t, Eq t, Typeable t, HasCallStack)

⇒ Trip t t

→ Version
From Version

→ Version
To Version

→ t

→ Expectation

Verify that round triping through the binary form holds for a range of versions.

In other words check that:

deserialize version . serialize version === id
serialize version . deserialize version . serialize version === serialize version

roundTripFailureExpectation ∷ (Typeable t, Eq t, HasCallStack) ⇒ Trip t t → t → Expectation

roundTripRangeFailureExpectation

Arguments

∷ ∀ t. (Typeable t, Eq t, HasCallStack)

⇒ Trip t t

→ Version
From Version

→ Version
To Version

→ t

→ Expectation

Enc/DecCBOR

roundTripCborExpectation ∷ ∀ t. (Show t, Eq t, EncCBOR t, DecCBOR t, HasCallStack) ⇒ t → Expectation

roundTripCborRangeExpectation

Arguments

∷ ∀ t. (Show t, Eq t, EncCBOR t, DecCBOR t, HasCallStack)

⇒ Version
From Version

→ Version
To Version

→ t

→ Expectation

roundTripCborFailureExpectation ∷ ∀ t. (EncCBOR t, DecCBOR t, Eq t, HasCallStack) ⇒ t → Expectation

roundTripCborRangeFailureExpectation

Arguments

∷ ∀ t. (EncCBOR t, DecCBOR t, Eq t, HasCallStack)

⇒ Version
From Version

→ Version
To Version

→ t

→ Expectation

roundTripAnnExpectation ∷ (Show t, Eq t, ToCBOR t, DecCBOR (Annotator t), HasCallStack) ⇒ t → Expectation

roundTripAnnRangeExpectation ∷ (Show t, Eq t, ToCBOR t, DecCBOR (Annotator t), HasCallStack) ⇒ Version → Version → t → Expectation

roundTripAnnFailureExpectation ∷ (ToCBOR t, DecCBOR (Annotator t), HasCallStack) ⇒ t → Expectation

roundTripAnnRangeFailureExpectation ∷ (ToCBOR t, DecCBOR (Annotator t), HasCallStack) ⇒ Version → Version → t → Expectation

Embed

embedTripSpec

Arguments

∷ ∀ a b. (Show a, Typeable a, Typeable b, Arbitrary a, Eq b, HasCallStack)

⇒ Version
Version for the encoder

→ Version
Version for the decoder

→ Trip a b

→ (b → a → Expectation)

→ Spec

Tests the embedtrip property using QuickCheck generators

embedTripExpectation

Arguments

∷ ∀ a b. (Typeable b, Eq b, HasCallStack)

⇒ Version
Version for the encoder

→ Version
Version for the decoder

→ Trip a b

→ (b → a → Expectation)

→ a

→ Expectation

embedTripAnnExpectation

Arguments

∷ ∀ a b. (ToCBOR a, DecCBOR (Annotator b), HasCallStack)

⇒ Version
Version for the encoder

→ Version
Version for the decoder

→ (b → a → Expectation)

→ a

→ Expectation

This is just like roundTripAnnExpectation, except it allows for source and target types to be different. This is very useful to test translation of the same type family from one era to another.

embedTripFailureExpectation ∷ (Typeable b, Eq b, HasCallStack) ⇒ Trip a b → a → Expectation

embedTripRangeFailureExpectation

Arguments

∷ ∀ a b. (Typeable b, Eq b, HasCallStack)

⇒ Trip a b

→ Version
From Version

→ Version
To Version

→ a

→ Expectation

roundTripTwiddledProperty ∷ (Show t, Eq t, Twiddle t, DecCBOR t) ⇒ Version → t → Property

roundTripAnnTwiddledProperty ∷ ∀ t q. (Twiddle t, DecCBOR (Annotator t), Testable q) ⇒ (t → t → q) → Version → t → Property

Tripping failure

data RoundTripFailure

Constructors

RoundTripFailure

Fields

rtfEncoderVersion ∷ Version

Version that was used during encoding

rtfDecoderVersion ∷ Version

Version that was used during decoding

rtfEncoding ∷ Encoding

Produced plain encoding

rtfEncodedBytes ∷ ByteString

Serialized encoding using the version in this failure

rtfReEncodedBytes ∷ Maybe ByteString

Re-serialized bytes, if there was a mismatch between the binary form and the reserialization of the data type.

rtfConformanceError ∷ Maybe String

Roundtripping through FlatTerm

rtfDropperError ∷ Maybe DecoderError

Error received while decoding the produced bytes and dropping the value. Normally it will be Nothing, unless the error produced did not match the rtfDecoderError, in which case it will be Just the error.

rtfDecoderError ∷ Maybe DecoderError

Error received while decoding the produced bytes. It is possible for a dropper to produce an error, while decoder going through successfully, which constitues a test failure. In such a case this field will be Nothing, however rtfDropperError will be set to Just. Whenever both rtfDropperError and rtfDecoderError are Nothing it means that the decoding went though just fine, but there was a mismatch in the binary format, i.e. reserialization produced a mismatched result, in which case rtfReEncodedBytes will be set to Just

Tripping definitions

data Trip a b

A definition of a CBOR trip through binary representation of one type to another. In this module this is called an embed. When a source and target type is the exact same one then it would be a dual and is expected to round trip.

Constructors

Trip

Fields

tripEncoder ∷ a → Encoding

tripDecoder ∷ ∀ s. Decoder s b

tripDropper ∷ ∀ s. Decoder s ()

mkTrip ∷ ∀ a b. (a → Encoding) → (∀ s. Decoder s b) → Trip a b

Construct a Trip using encoder and decoder, with dropper set to the decoder which drops the value

cborTrip ∷ ∀ a b. (EncCBOR a, DecCBOR b) ⇒ Trip a b

Tripping functions

roundTrip ∷ ∀ t. (Eq t, Typeable t) ⇒ Version → Trip t t → t → Either RoundTripFailure t

Check that serialization followed by deserialization of the value produces the same value back. We also check that re-serialization is idempotent. In other words, we ensure that deserialization does not modify the decoded value in a way that its binary representation has changed. Dropper is checked as well.

roundTripTwiddled ∷ ∀ t. (Twiddle t, DecCBOR t, Eq t) ⇒ Version → t → Gen (Either RoundTripFailure t)

roundTripAnn ∷ (ToCBOR t, DecCBOR (Annotator t)) ⇒ Version → t → Either RoundTripFailure t

roundTripAnnTwiddled ∷ (Twiddle t, DecCBOR (Annotator t)) ⇒ Version → t → Gen (Either RoundTripFailure t)

embedTrip

Arguments

∷ ∀ a b. (Eq b, Typeable b)

⇒ Version
Version for the encoder

→ Version
Version for the decoder

→ Trip a b

→ a

→ Either RoundTripFailure b

Can we serialise a type, and then deserialise it as something else?

embedTripAnn

Arguments

∷ ∀ a b. (ToCBOR a, DecCBOR (Annotator b))

⇒ Version
Encoder version for test failure reporting

→ Version
Decoder version

→ a

→ Either RoundTripFailure b

embedTripLabel

Arguments

∷ ∀ a b. Eq b

⇒ Text

→ Version
Version for the encoder

→ Version
Version for the decoder

→ Trip a b

→ a

→ Either RoundTripFailure b

embedTripLabelExtra

Arguments

∷ ∀ a b. Eq b

⇒ Text

→ Version
Version for the encoder

→ Version
Version for the decoder

→ Trip a b

→ a

→ Either RoundTripFailure (b, Encoding, ByteString)

decodeAnnExtra

Arguments

∷ ∀ t. Text

→ Version
Version for the encoder

→ Version
Version for the decoder

→ (∀ s. Decoder s (Annotator t))

→ Encoding

→ Either RoundTripFailure (t, ByteString)