Test.Cardano.Ledger.Binary.TreeDiff

Documentation

class ToExpr a where

toExpr converts a Haskell value into untyped Haskell-like syntax tree, Expr.

>>> toExpr ((1, Just 2) :: (Int, Maybe Int))
App "_\215_" [App "1" [],App "Just" [App "2" []]]

Minimal complete definition

Nothing

Methods

toExpr ∷ a → Expr

listToExpr ∷ [a] → Expr

newtype CBORBytes

Constructors

CBORBytes

Fields

unCBORBytes ∷ ByteString

newtype HexBytes

Wraps regular ByteString, but shows and diffs it as hex

Constructors

HexBytes

Fields

unHexBytes ∷ ByteString

showExpr ∷ ToExpr a ⇒ a → String

ansiExpr ∷ ToExpr a ⇒ a → Doc AnsiStyle

ansiExprString ∷ ToExpr a ⇒ a → String

diffExpr ∷ ToExpr a ⇒ a → a → Doc AnsiStyle

diffExprString ∷ ToExpr a ⇒ a → a → String

diffExprCompact ∷ ToExpr a ⇒ a → a → Doc AnsiStyle

diffExprCompactString ∷ ToExpr a ⇒ a → a → String

ansiDocToString ∷ Doc AnsiStyle → String

hexByteStringExpr ∷ ByteString → [Expr]

showHexBytesGrouped ∷ Int → ByteString → [String]

Show a ByteString as hex groups of 8bytes each. This is a slightly more useful form for debugging, rather than bunch of escaped characters.

assertColorFailure ∷ HasCallStack ⇒ String → IO a

Similar to assertFailure, except hspec will not interfer with any escape sequences that indicate color output.

expectExprEqual ∷ (Eq a, ToExpr a) ⇒ a → a → Expectation

Check that two values are equal and if they are not raise an exception with the ToExpr diff

expectExprEqualWithMessage ∷ (ToExpr a, Eq a, HasCallStack) ⇒ String → a → a → Expectation

Use this with HSpec, but with Tasty use assertExprEqualWithMessage below

assertExprEqualWithMessage ∷ (ToExpr a, Eq a, HasCallStack) ⇒ String → a → a → Assertion

Use this with Tasty, but with HSpec use expectExprEqualWithMessage above

callStackToLocation ∷ CallStack → Maybe Location

srcLocToLocation ∷ SrcLoc → Location

data Expr

A untyped Haskell-like expression.

Having richer structure than just Tree allows to have richer diffs.

Constructors

App ConstructorName [Expr]
application

Rec ConstructorName (OMap FieldName Expr)
record constructor

Lst [Expr]
list constructor

defaultExprViaShow ∷ Show a ⇒ a → Expr

An alternative implementation for literal types. We use show representation of them.

trimExprViaShow ∷ Show a ⇒ Int → a → Expr

tableDoc ∷ Maybe (Doc AnsiStyle) → [(String, Doc AnsiStyle)] → Doc AnsiStyle

data Pretty doc

Because we don't want to commit to single pretty printing library, we use explicit dictionary.

Constructors

Pretty

Fields

ppCon ∷ ConstructorName → doc

Display ConstructorName

ppApp ∷ doc → [doc] → doc

Display App

ppRec ∷ doc → [(FieldName, doc)] → doc

Display Rec

ppLst ∷ [doc] → doc

Display Lst

ppCpy ∷ doc → doc

Display unchanged parts

ppIns ∷ doc → doc

Display added parts

ppDel ∷ doc → doc

Display removed parts

ppEdits ∷ [doc] → doc

Combined edits (usually some sep combinator)

ppEllip ∷ doc

Ellipsis

ppParens ∷ doc → doc

Parens an expression

data Doc ann

The abstract data type Doc ann represents pretty documents that have been annotated with data of type ann.

More specifically, a value of type Doc represents a non-empty set of possible layouts of a document. The layout functions select one of these possibilities, taking into account things like the width of the output document.

The annotation is an arbitrary piece of data associated with (part of) a document. Annotations may be used by the rendering backends in order to display output differently, such as

color information (e.g. when rendering to the terminal)

mouseover text (e.g. when rendering to rich HTML)

whether to show something or not (to allow simple or detailed versions)

The simplest way to display a Doc is via the Show class.

>>> putStrLn (show (vsep ["hello", "world"]))
hello
world

data AnsiStyle

Render the annotated document in a certain style. Styles not set in the annotation will use the style of the surrounding document, or the terminal’s default if none has been set yet.

style = color Green <> bold
styledDoc = annotate style "hello world"

ansiWlPretty ∷ Pretty Doc

Pretty via ansi-wl-pprint library (with colors).

ppEditExpr ∷ Pretty doc → Edit EditExpr → doc

Pretty print an Edit EditExpr using explicit pretty-printing dictionary.

ediff ∷ ToExpr a ⇒ a → a → Edit EditExpr

Difference between two ToExpr values.

>>> let x = (1, Just 2) :: (Int, Maybe Int)
>>> let y = (1, Nothing)
>>> prettyEditExpr (ediff x y)
_×_ 1 -(Just 2) +Nothing

>>> data Foo = Foo { fooInt :: Either Char Int, fooBool :: [Maybe Bool], fooString :: String } deriving (Eq, Generic)
>>> instance ToExpr Foo

>>> prettyEditExpr $ ediff (Foo (Right 2) [Just True] "fo") (Foo (Right 3) [Just True] "fo")
Foo {fooInt = Right -2 +3, fooBool = [Just True], fooString = "fo"}

>>> prettyEditExpr $ ediff (Foo (Right 42) [Just True, Just False] "old") (Foo (Right 42) [Nothing, Just False, Just True] "new")
Foo {
  fooInt = Right 42,
  fooBool = [-Just True, +Nothing, Just False, +Just True],
  fooString = -"old" +"new"}

Orphan instances

ToExpr Version

Instance details

Methods

toExpr ∷ Version → Expr

listToExpr ∷ [Version] → Expr

ToExpr Term

Instance details

Methods

toExpr ∷ Term → Expr

listToExpr ∷ [Term] → Expr

ToExpr IPv4

Instance details

Methods

toExpr ∷ IPv4 → Expr

listToExpr ∷ [IPv4] → Expr

ToExpr IPv6

Instance details

Methods

toExpr ∷ IPv6 → Expr

listToExpr ∷ [IPv6] → Expr

ToExpr a ⇒ ToExpr (Sized a)

Instance details

Methods

toExpr ∷ Sized a → Expr

listToExpr ∷ [Sized a] → Expr

ToExpr a ⇒ ToExpr (StrictMaybe a)

Instance details

Methods

toExpr ∷ StrictMaybe a → Expr

listToExpr ∷ [StrictMaybe a] → Expr

ToExpr a ⇒ ToExpr (StrictSeq a)

Instance details

Methods

toExpr ∷ StrictSeq a → Expr

listToExpr ∷ [StrictSeq a] → Expr

DSIGNAlgorithm c ⇒ ToExpr (SignedDSIGN c index)

Instance details

Methods

toExpr ∷ SignedDSIGN c index → Expr

listToExpr ∷ [SignedDSIGN c index] → Expr

ToExpr (Hash c index)

Instance details

Methods

toExpr ∷ Hash c index → Expr

listToExpr ∷ [Hash c index] → Expr