Test.Cardano.Ledger.Binary.Twiddle

Documentation

data Twiddler a

Constructors

Twiddler

Fields

twiddlerVersion ∷ !Version

twiddlerType ∷ !a

twiddlerTerm ∷ Term

class Twiddle a where

Introducing random variations into the CBOR encoding of a value while preserving the round-trip properties.

For any value `x :: a`, where a derives Twiddle, and for any version of the decoder, the following property must hold: >>> fmap ((== x) . encodingToTerm version . encCBOR) (twiddle x)

Minimal complete definition

Nothing

Methods

twiddle ∷ Version → a → Gen Term

Given a value of type a, generates a CBOR Term that can contain slight variations without changing the semantics. After encoding and decoding the Term, we should get back the original value that was being twiddled.

In addition to varying the low-level CBOR tokens, twiddle can also be used to introduce higher level variations. For example if the schema of a value allows a field to be either an empty list or absent entirely, and both are interpreted the same way, then twiddle can be used to randomly pick either of these representations.

default twiddle ∷ ∀ p. (Generic a, TwiddleL (Rep a p)) ⇒ Version → a → Gen Term

encodingToTerm ∷ Version → Encoding → Term

toTwiddler ∷ Twiddle a ⇒ Version → a → Gen (Twiddler a)

Wraps an arbitrary value into a Twiddler

toTerm ∷ EncCBOR a ⇒ Version → a → Term

twiddleInvariantProp ∷ ∀ a. Twiddle a ⇒ Version → a → Gen Property

Function for testing the invariant of a Twiddle instance. For a correct implementation, this property should always hold.

emptyOrNothing ∷ ∀ t b. (Foldable t, Twiddle (t Void), Monoid (t Void), Twiddle (t b)) ⇒ Version → t b → Gen (Maybe Term)

Optional containers have two "empty" representations. One of them is to return an empty container and the other is to omit the field. This utility function randomly picks one of these representations, where omission is represented by Nothing and empty container is returned with Just. These values can then be easily concatenated with catMaybes.

twiddleStrictMaybe ∷ Twiddle a ⇒ Version → StrictMaybe a → Gen (Maybe Term)

Utility function for twiddling optional fields. It works similarly to the twiddle method of StrictMaybe, but lifts the Maybe constructor out from the term so it can be easily concatenated with catMaybes.