Test.Cardano.Ledger.Binary.Vintage.Helpers

Documentation

byronProtVer ∷ Version

Binary test helpers

data U

data U24

Like U, but we expect to read back the Cbor Data Item when decoding.

extensionProperty ∷ ∀ a. (Arbitrary a, Eq a, Show a, DecCBOR a, EncCBOR a) ⇒ Property

Given a data type which can be extended, verify we can indeed do so without breaking anything. This should work with every time which adopted the schema of having at least one constructor of the form: .... | Unknown Word8 ByteString

cborFlatTermValid ∷ EncCBOR a ⇒ a → Property

Machinery to test we perform "flat" encoding.

Static size estimates

data SizeTestConfig a

Configuration for a single test case.

Constructors

SizeTestConfig

Fields

debug ∷ a → String

Pretty-print values

gen ∷ Gen a

Generator

precise ∷ Bool

Must estimates be exact?

addlCtx ∷ Map TypeRep SizeOverride

Additional size overrides

computedCtx ∷ a → Map TypeRep SizeOverride

Size overrides computed from a concrete instance.

cfg ∷ Buildable a ⇒ SizeTestConfig a

Default configuration, for Buildable types.

scfg ∷ Show a ⇒ SizeTestConfig a

Default configuration, for Showable types.

sizeTest ∷ ∀ a. EncCBOR a ⇒ SizeTestConfig a → Property

Create a test case from the given test configuration.