Cardano.Ledger.Binary.Decoding.DecCBOR

Documentation

class Typeable a ⇒ DecCBOR a where

Minimal complete definition

Nothing

Methods

decCBOR ∷ Decoder s a

default decCBOR ∷ FromCBOR a ⇒ Decoder s a

dropCBOR ∷ Proxy a → Decoder s ()

Validate decoding of a Haskell value, without the need to actually construct it. Coule be slightly faster than decCBOR, however it should respect this law:

dropCBOR (proxy :: Proxy a) = () <$ (decCBOR :: Decoder s a)

label ∷ Proxy a → Text

fromByronCBOR ∷ DecCBOR a ⇒ Decoder s a

Convert a versioned DecCBOR instance to a plain Decoder using Byron protocol version.