Cardano.Ledger.Binary.FlatTerm

Documentation

type FlatTerm = [TermToken]

A "flat" representation of an Encoding value, useful for round-tripping and writing tests.

Since: cborg-0.2.0.0

data TermToken

A concrete encoding of Encoding values, one which mirrors the original Encoding type closely.

Since: cborg-0.2.0.0

Constructors

TkInt !Int

TkInteger !Integer

TkBytes !ByteString

TkBytesBegin

TkString !Text

TkStringBegin

TkListLen !Word

TkListBegin

TkMapLen !Word

TkMapBegin

TkBreak

TkTag !Word64

TkBool !Bool

TkNull

TkSimple !Word8

TkFloat16 !Float

TkFloat32 !Float

TkFloat64 !Double

toFlatTerm ∷ Version → Encoding → FlatTerm

fromFlatTerm ∷ Version → (∀ s. Decoder s a) → FlatTerm → Either String a

validFlatTerm

Arguments

∷ FlatTerm
The input FlatTerm

→ Bool
True if valid, False otherwise.

Ensure a FlatTerm is internally consistent and was created in a valid manner.

Since: cborg-0.2.0.0