Test.Cardano.Ledger.Binary.TreeDiff

Documentation

newtype CBORBytes

Constructors

CBORBytes

Fields

unCBORBytes ∷ ByteString

newtype HexBytes

Wraps regular ByteString, but shows and diffs it as hex

Constructors

HexBytes

Fields

unHexBytes ∷ ByteString

showExpr ∷ ToExpr a ⇒ a → String

diffExpr ∷ ToExpr a ⇒ a → a → String

hexByteStringExpr ∷ ByteString → [Expr]

showHexBytesGrouped ∷ ByteString → [String]

Show a ByteString as hex groups of 8bytes each. This is a slightly more useful form for debugging, rather than bunch of escaped characters.

expectExprEqual ∷ (Eq a, ToExpr a) ⇒ a → a → Expectation

Check that two values are equal and if they are not raise an exception with the ToExpr diff

expectExprEqualWithMessage ∷ (ToExpr a, Eq a, HasCallStack) ⇒ [Char] → a → a → Expectation

Use this with HSpec, but with Tasty use assertExprEqualWithMessage below

assertExprEqualWithMessage ∷ (ToExpr a, Eq a, HasCallStack) ⇒ [Char] → a → a → Assertion

Use this with Tasty, but with HSpec use expectExprEqualWithMessage above

Orphan instances

ToExpr Term

Instance details

Methods

toExpr ∷ Term → Expr

listToExpr ∷ [Term] → Expr